+++ packages/runtimed/src/index.ts
@@ -149,20 +149,28 @@ export {
 // Derived state
 export {
   type DaemonQueueState,
+  deriveEnvironmentYml,
   deriveEnvManager,
   deriveEnvSyncState,
   deriveKernelInfo,
+  derivePixiInfo,
+  derivePyproject,
   deriveQueueState,
   deriveRuntimeKind,
   type EnvManager,
   type EnvManagerMetadataInputs,
   type EnvSyncDiff,
   type EnvSyncState,
+  type EnvironmentYmlDeps,
+  type EnvironmentYmlInfo,
   isKernelStatus,
   KERNEL_STATUS,
   type KernelInfo,
   type KernelStatus,
   lifecycleToLegacyStatus,
+  type PixiInfo,
+  type PyProjectDeps,
+  type PyProjectInfo,
   RUNTIME_STATUS,
   runtimeStatusKey,
   type RuntimeKind,
